Reorganizes cleaning tasks into a task module and maintenance DAGs into a suitable directory.
